refactor(desktop): split memory atlas responsibilities
