+++ desktop/macos/Desktop/Sources/MainWindow/Pages/MemoryGraph/CanonicalMemoryAtlasView.swift
@@ -34,1726 +34,6 @@ extension Notification.Name {
   )
 }
 
-// MARK: - Atlas Layout
-
-enum MemoryAtlasCluster: String, CaseIterable, Identifiable {
-  case person
-  case organization
-  case place
-  case thing
-  case concept
-
-  var id: String { rawValue }
-
-  var title: String {
-    switch self {
-    case .person: return "People"
-    case .organization: return "Organizations"
-    case .place: return "Places"
-    case .thing: return "Things"
-    case .concept: return "Concepts"
-    }
-  }
-
-  var color: Color {
-    switch self {
-    case .person: return Color(nsColor: .systemTeal)
-    case .organization: return Color(nsColor: .systemOrange)
-    case .place: return Color(nsColor: .systemGreen)
-    case .thing: return Ink.secondary
-    case .concept: return Color(nsColor: .systemBlue)
-    }
-  }
-
-  /// Present types orbit the person at the center as a shallow five-petal
-  /// constellation. The two radii counterbalance the wide desktop canvas so
-  /// the ring reads as a circle rather than a flattened row.
-  static let starCenter = CGPoint(x: 0.5, y: 0.5)
-  private static let starHorizontalRadius: CGFloat = 0.15
-  private static let starVerticalRadius: CGFloat = 0.25
-
-  static func centers(for activeClusters: [Self]) -> [Self: CGPoint] {
-    guard !activeClusters.isEmpty else { return [:] }
-
-    // A single active type has no constellation to form. Parking it on the ring
-    // would strand every node in one corner of the canvas and leave the rest
-    // empty, so it takes the center and orbits the anchor directly.
-    if activeClusters.count == 1 {
-      return [activeClusters[0]: starCenter]
-    }
-
-    let startingAngle = -Double.pi / 2
-    let angularStep = 2 * Double.pi / Double(activeClusters.count)
-    var result: [Self: CGPoint] = [:]
-    for (index, cluster) in activeClusters.enumerated() {
-      let angle = startingAngle + Double(index) * angularStep
-      result[cluster] = CGPoint(
-        x: starCenter.x + CGFloat(cos(angle)) * starHorizontalRadius,
-        y: starCenter.y + CGFloat(sin(angle)) * starVerticalRadius
-      )
-    }
-    return result
-  }
-}
-
-struct MemoryAtlasEdgePlacement: Identifiable {
-  let edge: KnowledgeGraphEdge
-  let source: CGPoint
-  let target: CGPoint
-  let cluster: MemoryAtlasCluster
-  /// Distinct relationship labels merged from parallel server edges between
-  /// this pair (same endpoints, different verbs), in first-seen order. `edge`
-  /// carries the first label for callers that only read a single verb;
-  /// `MemoryAtlasLayoutEngine.combinedRelationshipDisplayName` renders all of
-  /// them together for the inspector.
-  let relationshipLabels: [String]
-  /// Count of underlying server edges collapsed into this placement.
-  /// Deliberately unused by rendering today — preserved as spring-strength
-  /// input for a future force-directed layout.
-  let weight: Int
-  /// Whether both ends of this connection sit in the same neighbourhood.
-  ///
-  /// What lets the wiring itself show the grouping: drawn a little brighter
-  /// inside a region than between two, the mesh thickens where a group is
-  /// dense and thins at its edges, so the structure is legible without
-  /// reading the territory layer at all.
-  let withinNeighbourhood: Bool
-
-  var id: String { edge.id }
-}
-
-/// A region of the map the layout found by structure rather than by type.
-///
-/// The relaxation already has to detect these to draw them apart. Naming them
-/// is what turns "the dots are arranged meaningfully" into something a person
-/// can act on: a place they can see from across the map and decide to enter.
-struct MemoryAtlasNeighbourhood: Identifiable, Equatable {
-  /// Opaque. Group numbers come out of a modularity pass and carry no meaning
-  /// across rebuilds, so nothing user-facing may be derived from this.
-  let id: Int
-  let memberIDs: [String]
-  /// Its most-connected few members, named.
-  ///
-  /// Deliberately not a generated title. A model asked to name this group
-  /// would happily answer "Career" or "Family", and be confidently wrong in a
-  /// way the user cannot check. Listing who is actually in it says exactly what
-  /// the algorithm did and nothing more, and is falsifiable at a glance.
-  let caption: String
-  /// Normalized centre of its members.
-  let center: CGPoint
-  /// Normalized radius covering most of them. No longer the drawn shape — the
-  /// coastline is — but still what decides where a caption goes and how far the
-  /// camera pulls back when the region is entered.
-  let radius: CGFloat
-  /// The territory this neighbourhood holds, as closed rings in normalized map
-  /// coordinates. Usually one; more than one when the group lives in two places
-  /// and the map declines to invent a land bridge between them.
-  let coastline: [[CGPoint]]
-  /// Of the entities standing on this territory, the share that belong to it.
-  ///
-  /// Whether the region is a place or merely an average. A neighbourhood can be
-  /// perfectly well detected and still be smeared through two others, and the
-  /// difference is invisible in the caption: both are a list of names over a
-  /// patch of canvas. Kept as the map's own check on the shape it drew, and
-  /// measured against the coastline rather than a disc so it describes the
-  /// ground actually claimed.
-  let purity: Double
-
-  var memberCount: Int { memberIDs.count }
-}
-
-struct MemoryAtlasSnapshot {
-  let nodes: [MemoryAtlasNodePlacement]
-  let edges: [MemoryAtlasEdgePlacement]
-  let anchorNodeID: String?
-  /// Largest first. Empty on maps too small or too sparse to have regions.
-  let neighbourhoods: [MemoryAtlasNeighbourhood]
-  let activeClusters: [MemoryAtlasCluster]
-  let clusterCenters: [MemoryAtlasCluster: CGPoint]
-  let nodeByID: [String: MemoryAtlasNodePlacement]
-  /// Edge order is computed once when the graph is received. Camera updates can
-  /// then filter this stable order instead of sorting the whole graph per frame.
-  let rankedEdges: [MemoryAtlasEdgePlacement]
-  let overviewEdges: [MemoryAtlasEdgePlacement]
-  let neighborhoodEdges: [MemoryAtlasEdgePlacement]
-  let detailEdges: [MemoryAtlasEdgePlacement]
-  let edgesByNodeID: [String: [MemoryAtlasEdgePlacement]]
-  let neighborIDsByNodeID: [String: Set<String>]
-  /// The time axis for this graph, or `nil` when timestamps carry no spread.
-  let timeline: MemoryAtlasTimeline?
-
-  init(
-    nodes: [MemoryAtlasNodePlacement],
-    edges: [MemoryAtlasEdgePlacement],
-    anchorNodeID: String?,
-    clusterCenters: [MemoryAtlasCluster: CGPoint],
-    neighbourhoods: [MemoryAtlasNeighbourhood] = []
-  ) {
-    self.nodes = nodes
-    self.edges = edges
-    self.anchorNodeID = anchorNodeID
-    self.neighbourhoods = neighbourhoods
-    self.activeClusters = MemoryAtlasCluster.allCases.filter { clusterCenters[$0] != nil }
-    self.clusterCenters = clusterCenters
-    self.timeline = MemoryAtlasTimeline.make(from: nodes.map(\.node))
-    let indexedNodes = Dictionary(lastWriteWins: nodes.map { ($0.id, $0) })
-    nodeByID = indexedNodes
-
-    let keptSpokes = MemoryAtlasSnapshot.spokesWorthDrawing(
-      edges, nodes: indexedNodes, anchorID: anchorNodeID)
-    let sortedEdges = edges.sorted { lhs, rhs in
-      let lhsRank = MemoryAtlasSnapshot.edgeRank(
-        lhs, nodes: indexedNodes, anchorID: anchorNodeID, keptSpokes: keptSpokes)
-      let rhsRank = MemoryAtlasSnapshot.edgeRank(
-        rhs, nodes: indexedNodes, anchorID: anchorNodeID, keptSpokes: keptSpokes)
-      return lhsRank < rhsRank
-    }
-    rankedEdges = sortedEdges
-
-    overviewEdges = sortedEdges.filter { edge in
-      MemoryAtlasSnapshot.maximumEndpointRank(edge, nodes: indexedNodes) < 4
-    }
-    neighborhoodEdges = sortedEdges.filter { edge in
-      MemoryAtlasSnapshot.maximumEndpointRank(edge, nodes: indexedNodes) < 8
-    }
-    detailEdges = sortedEdges.filter { edge in
-      MemoryAtlasSnapshot.maximumEndpointRank(edge, nodes: indexedNodes) < 12
-    }
-
-    var indexedEdges: [String: [MemoryAtlasEdgePlacement]] = [:]
-    var indexedNeighbors: [String: Set<String>] = [:]
-    for placement in sortedEdges {
-      let sourceID = placement.edge.sourceId
-      let targetID = placement.edge.targetId
-      indexedEdges[sourceID, default: []].append(placement)
-      indexedEdges[targetID, default: []].append(placement)
-      indexedNeighbors[sourceID, default: []].insert(targetID)
-      indexedNeighbors[targetID, default: []].insert(sourceID)
-    }
-    edgesByNodeID = indexedEdges
-    neighborIDsByNodeID = indexedNeighbors
-  }
-
-  func rankedEdges(for detailLevel: MemoryAtlasDetailLevel) -> [MemoryAtlasEdgePlacement] {
-    switch detailLevel {
-    case .overview: overviewEdges
-    case .neighborhood: neighborhoodEdges
-    case .detail: detailEdges
-    case .focus: detailEdges
-    case .inspect: detailEdges
-    }
-  }
-
-  func center(for cluster: MemoryAtlasCluster) -> CGPoint {
-    clusterCenters[cluster] ?? MemoryAtlasCluster.starCenter
-  }
-
-  /// What gets drawn first when there is only room for a few dozen lines.
-  ///
-  /// The account holder's own connections sort last, however prominent their
-  /// endpoints are. They used to sort *first* — the anchor's rank is zero, so
-  /// every edge touching it led the order — and the budget is small enough
-  /// (36 lines at overview) that on a real account it was spent entirely on
-  /// them. The map reported 1,377 connections and drew a hundred copies of the
-  /// one fact the user already knows: that everything here is theirs. The
-  /// relationships between two other entities, which are the only ones that
-  /// can tell them something, never made the cut.
-  /// How many of the account holder's own connections still compete for the
-  /// budget on merit.
-  ///
-  /// Not zero. Sending *every* spoke to the back leaves the one entity the user
-  /// is certain is connected to everything drawn with no lines at all — the
-  /// opposite lie from the starburst, and just as wrong. A handful, to their
-  /// most prominent partners, says "you are in this" without saying it a
-  /// hundred times.
-  static let spokesDrawnOnMerit = 6
-
-  static func spokesWorthDrawing(
-    _ edges: [MemoryAtlasEdgePlacement],
-    nodes: [String: MemoryAtlasNodePlacement],
-    anchorID: String?
-  ) -> Set<String> {
-    guard let anchorID else { return [] }
-    let partnerRank: (MemoryAtlasEdgePlacement) -> Int = { placement in
-      let other =
-        placement.edge.sourceId == anchorID ? placement.edge.targetId : placement.edge.sourceId
-      return nodes[other]?.clusterRank ?? .max
-    }
-    return Set(
-      edges
-        .filter { $0.edge.sourceId == anchorID || $0.edge.targetId == anchorID }
-        .sorted {
-          partnerRank($0) == partnerRank($1) ? $0.id < $1.id : partnerRank($0) < partnerRank($1)
-        }
-        .prefix(spokesDrawnOnMerit)
-        .map(\.id))
-  }
-
-  /// What gets drawn first when there is only room for a few dozen lines.
-  ///
-  /// The account holder's connections sort last beyond the first few, however
-  /// prominent their endpoints are. They used to sort *first* — the anchor's
-  /// rank is zero, so every edge touching it led the order — and the budget is
-  /// small enough (36 lines at overview) that on a real account it was spent
-  /// entirely on them. The map reported 1,377 connections and drew a hundred
-  /// copies of the one fact the user already knows: that everything here is
-  /// theirs. The relationships between two other entities, which are the only
-  /// ones that can tell them something, never made the cut.
-  static func edgeRank(
-    _ placement: MemoryAtlasEdgePlacement,
-    nodes: [String: MemoryAtlasNodePlacement],
-    anchorID: String?,
-    keptSpokes: Set<String> = []
-  ) -> (Int, Int, Int, String) {
-    let sourceRank = nodes[placement.edge.sourceId]?.clusterRank ?? .max
-    let targetRank = nodes[placement.edge.targetId]?.clusterRank ?? .max
-    let demoted =
-      anchorID != nil
-      && (placement.edge.sourceId == anchorID || placement.edge.targetId == anchorID)
-      && !keptSpokes.contains(placement.id)
-    return (demoted ? 1 : 0, min(sourceRank, targetRank), max(sourceRank, targetRank), placement.id)
-  }
-
-  private static func maximumEndpointRank(
-    _ placement: MemoryAtlasEdgePlacement,
-    nodes: [String: MemoryAtlasNodePlacement]
-  ) -> Int {
-    max(
-      nodes[placement.edge.sourceId]?.clusterRank ?? .max,
-      nodes[placement.edge.targetId]?.clusterRank ?? .max
-    )
-  }
-}
-
-enum MemoryAtlasDetailLevel: Equatable {
-  case overview
-  case neighborhood
-  case detail
-  case focus
-  case inspect
-}
-
-struct MemoryAtlasRenderPlan {
-  let visibleNodes: [MemoryAtlasNodePlacement]
-  let visibleEdges: [MemoryAtlasEdgePlacement]
-  let interactiveNodes: [MemoryAtlasNodePlacement]
-  let labelNodeIDs: Set<String>
-  /// Entities whose name is drawn above the mark instead of below it. Labels
-  /// default to hanging below; flipping is how a sparse atlas keeps a name that
-  /// would otherwise collide with an already-placed one and be dropped.
-  let labelAboveNodeIDs: Set<String>
-  /// Canvas labels from the automatic inspection threshold onward. Keeping
-  /// these outside the SwiftUI overlay means a large graph can label every
-  /// on-screen dot without building thousands of view/hit-test nodes per
-  /// frame.
-  let canvasLabelNodes: [MemoryAtlasNodePlacement]
-  let usesCanvasLabels: Bool
-  let isFullyLabelled: Bool
-  let relatedNodeIDs: Set<String>
-  let detailLevel: MemoryAtlasDetailLevel
-}
-
-enum MemoryAtlasRenderPlanner {
-  static func makePlan(
-    snapshot: MemoryAtlasSnapshot,
-    viewportSize: CGSize,
-    zoom: CGFloat,
-    pan: CGSize,
-    compact: Bool,
-    selectedNodeID: String?,
-    matchingNodeIDs: Set<String>?,
-    matchingEdges: [MemoryAtlasEdgePlacement]? = nil,
-    asOf: Date? = nil,
-    timeline: MemoryAtlasTimeline? = nil,
-    timeCursor: Double? = nil
-  ) -> MemoryAtlasRenderPlan {
-    let fullyLabelledZoom = MemoryAtlasZoomPolicy.fullyLabelledZoom(
-      nodeCount: snapshot.nodes.count
-    )
-    let isFullyLabelled = !compact && zoom >= fullyLabelledZoom
-    let usesCanvasLabels =
-      !compact
-      && zoom
-        >= MemoryAtlasZoomPolicy.automaticCanvasLabelZoom(
-          nodeCount: snapshot.nodes.count
-        )
-    let detailLevel: MemoryAtlasDetailLevel =
-      if zoom < MemoryAtlasZoomPolicy.neighborhoodZoom {
-        .overview
-      } else if zoom < 1.9 {
-        .neighborhood
-      } else if zoom < MemoryAtlasZoomPolicy.focusModeZoom {
-        .detail
-      } else if zoom < MemoryAtlasZoomPolicy.inspectModeZoom {
-        .focus
-      } else {
-        .inspect
-      }
-
-    // Detail must be additive. The previous planner reduced the node budget
-    // from 1,200 to 600 immediately after overview, which made visible dots
-    // disappear on a small zoom-in. Keep a stable, salience-ordered cohort and
-    // only add more of it as fidelity increases.
-    let maximumNodeLimit: Int =
-      if isFullyLabelled {
-        snapshot.nodes.count
-      } else {
-        switch detailLevel {
-        case .overview: 1_200
-        case .neighborhood: 1_600
-        case .detail: 2_400
-        case .focus, .inspect: 3_200
-        }
-      }
-    // The mesh is the structure, so it is not rationed at overview any more.
-    //
-    // Thirty-six of roughly fourteen hundred connections left the whole map to
-    // be read from the positions of dots alone: clustering was there, and
-    // invisible, because the thing that shows a group is dense wiring inside it
-    // and sparse wiring out of it. Average degree on a real account is under
-    // three, so the full mesh drawn as faint hairlines is a texture rather than
-    // a thicket. The node budget above still bounds how many endpoints exist.
-    let edgeLimit: Int =
-      switch detailLevel {
-      case .overview: 2_000
-      case .neighborhood: 2_400
-      case .detail: 3_000
-      case .focus: 3_600
-      case .inspect: 4_200
-      }
-    // These budgets exist to keep a multi-thousand-entity graph legible. An
-    // atlas small enough to name in full needs no rationing: withholding
-    // labels there just leaves unreadable dots on an empty canvas. Collision
-    // admission still decides what actually fits.
-    let isSmallAtlas = !compact && snapshot.nodes.count <= MemoryAtlasZoomPolicy.smallAtlasCeiling
-    let labelsPerCluster: Int =
-      if isSmallAtlas {
-        snapshot.nodes.count
-      } else {
-        switch detailLevel {
-        case .overview: compact ? 2 : 3
-        case .neighborhood: compact ? 4 : 7
-        case .detail: compact ? 5 : 11
-        case .focus: compact ? 5 : 24
-        case .inspect: compact ? 5 : 96
-        }
-      }
-    let labelLimit: Int =
-      if isSmallAtlas {
-        snapshot.nodes.count
-      } else {
-        switch detailLevel {
-        case .overview: 12
-        case .neighborhood: 24
-        case .detail: 36
-        case .focus: 72
-        case .inspect: 96
-        }
-      }
-
-    var relatedNodeIDs: Set<String> = []
-    if let selectedNodeID {
-      relatedNodeIDs = snapshot.neighborIDsByNodeID[selectedNodeID] ?? []
-      relatedNodeIDs.insert(selectedNodeID)
-    }
-
-    // The time cursor is a visibility filter layered over the stable layout: a
-    // node keeps its position and simply has not been "born" yet. The anchor is
-    // always present — "you" are the constant the rest of the memory accretes
-    // around.
-    let timeFilteredNodes: [MemoryAtlasNodePlacement]
-    if let timeline, let timeCursor, timeCursor < 0.9995 {
-      timeFilteredNodes = snapshot.nodes.filter { placement in
-        placement.id == snapshot.anchorNodeID || timeline.isVisible(nodeID: placement.id, at: timeCursor)
-      }
-    } else if let asOf {
-      timeFilteredNodes = snapshot.nodes.filter { placement in
-        placement.id == snapshot.anchorNodeID || placement.node.createdAt <= asOf
-      }
-    } else {
-      timeFilteredNodes = snapshot.nodes
-    }
-
-    // Camera movement changes where a node is painted, not whether it belongs
-    // to the rendered cohort. Canvas clipping handles off-screen content while
-    // this stable source order guarantees that zoom never drops entities just
-    // because a threshold or viewport candidate set changed.
-    let visibleNodes = priorityOrderedPrefix(
-      timeFilteredNodes,
-      limit: maximumNodeLimit,
-      anchorNodeID: snapshot.anchorNodeID,
-      selectedNodeID: selectedNodeID,
-      relatedNodeIDs: relatedNodeIDs,
-      matchingNodeIDs: matchingNodeIDs,
-      includeBackgroundNodes: true
-    )
-    let visibleNodeIDs = Set(visibleNodes.map(\.id))
-
-    let edgeCandidates: [MemoryAtlasEdgePlacement]
-    if let selectedNodeID {
-      edgeCandidates = snapshot.edgesByNodeID[selectedNodeID] ?? []
-    } else if let matchingNodeIDs {
-      edgeCandidates =
-        matchingEdges
-        ?? snapshot.rankedEdges.filter { edge in
-          matchingNodeIDs.contains(edge.edge.sourceId) || matchingNodeIDs.contains(edge.edge.targetId)
-        }
-    } else {
-      edgeCandidates = snapshot.rankedEdges
-    }
-
-    let selectedEdgeLimit = selectedNodeID == nil ? edgeLimit : min(edgeLimit, 80)
-    let visibleEdges = Array(
-      edgeCandidates.lazy
-        .filter { edge in
-          let isWithinTimeline =
-            timeline.flatMap { timeline in
-              timeCursor.map { cursor in
-                cursor >= 0.9995 || timeline.fraction(for: edge.edge.createdAt) <= cursor
-              }
-            } ?? true
-          let isBeforeAsOf = asOf.map { edge.edge.createdAt <= $0 } ?? true
-          return isWithinTimeline
-            && isBeforeAsOf
-            && visibleNodeIDs.contains(edge.edge.sourceId) && visibleNodeIDs.contains(edge.edge.targetId)
-        }
-        .prefix(selectedEdgeLimit)
-    )
-    let labelableNodes = visibleNodes.filter {
-      MemoryAtlasCatalogLayout.allowsAutomaticLabel(
-        isCatalog: $0.isCatalog, id: $0.id, selectedNodeID: selectedNodeID, matchingNodeIDs: matchingNodeIDs)
-    }
-    let labelCandidates: [MemoryAtlasNodePlacement]
-    if detailLevel == .inspect {
-      labelCandidates = labelableNodes
-    } else if selectedNodeID != nil {
-      labelCandidates = labelableNodes.filter { relatedNodeIDs.contains($0.id) }
-    } else if let matchingNodeIDs {
-      labelCandidates = labelableNodes.filter { matchingNodeIDs.contains($0.id) }
-    } else {
-      labelCandidates = labelableNodes.filter { placement in
-        placement.id == snapshot.anchorNodeID || placement.clusterRank < labelsPerCluster
-      }
-    }
-    let labels = admitLabels(
-      labelCandidates,
-      limit: labelLimit,
-      viewportSize: viewportSize,
-      zoom: zoom,
-      pan: pan,
-      compact: compact,
-      forcedNodeIDs: Set([selectedNodeID, snapshot.anchorNodeID].compactMap { $0 }),
-      allowsFlipAbove: isSmallAtlas
-    )
-
-    return MemoryAtlasRenderPlan(
-      visibleNodes: visibleNodes,
-      visibleEdges: visibleEdges,
-      // Every entity remains on the Canvas at deep zoom, while the expensive
-      // SwiftUI hit-target/label overlay stays bounded and grows gradually.
-      interactiveNodes: labels.placements,
-      // Once inspection has enough density-aware space, labels move into
-      // Canvas immediately. They therefore appear while panning/zooming and
-      // are not gated by selection or by the bounded SwiftUI overlay.
-      labelNodeIDs: usesCanvasLabels ? [] : Set(labels.placements.map(\.id)),
-      labelAboveNodeIDs: usesCanvasLabels ? [] : labels.aboveNodeIDs,
-      canvasLabelNodes: usesCanvasLabels ? labelableNodes : [],
-      usesCanvasLabels: usesCanvasLabels,
-      isFullyLabelled: isFullyLabelled,
-      relatedNodeIDs: relatedNodeIDs,
-      detailLevel: detailLevel
-    )
-  }
-
-  /// Fixed, non-interactive overview used by the Memories page. It keeps the
-  /// preview cheap even for large graphs and deliberately does no camera work.
-  static func makePreviewPlan(
-    snapshot: MemoryAtlasSnapshot,
-    nodeLimit: Int = 260,
-    edgeLimit: Int = 24
-  ) -> MemoryAtlasRenderPlan {
-    let visibleNodes = priorityOrderedPrefix(
-      snapshot.nodes,
-      limit: nodeLimit,
-      anchorNodeID: snapshot.anchorNodeID,
-      selectedNodeID: nil,
-      relatedNodeIDs: [],
-      matchingNodeIDs: nil,
-      includeBackgroundNodes: true
-    )
-    let visibleNodeIDs = Set(visibleNodes.map(\.id))
-    let visibleEdges = Array(
-      snapshot.overviewEdges.lazy
-        .filter {
-          visibleNodeIDs.contains($0.edge.sourceId) && visibleNodeIDs.contains($0.edge.targetId)
-        }
-        .prefix(edgeLimit)
-    )
-    return MemoryAtlasRenderPlan(
-      visibleNodes: visibleNodes,
-      visibleEdges: visibleEdges,
-      interactiveNodes: [],
-      labelNodeIDs: [],
-      labelAboveNodeIDs: [],
-      canvasLabelNodes: [],
-      usesCanvasLabels: false,
-      isFullyLabelled: false,
-      relatedNodeIDs: [],
-      detailLevel: .overview
-    )
-  }
-
-  static func renderedPoint(
-    for normalized: CGPoint,
-    viewportSize: CGSize,
-    zoom: CGFloat,
-    pan: CGSize
-  ) -> CGPoint {
-    // Must use the same square span as the drawing path's `point(for:in:)`:
-    // min(width, height) for both axes. Scaling x by the full viewport width
-    // and y by the full height made collision detection believe horizontally
-    // adjacent labels were farther apart than their rendered positions on
-    // wide desktop windows, admitting labels that overlapped on the canvas.
-    let span = MemoryAtlasLayoutEngine.projectionSpan(of: viewportSize)
-    return CGPoint(
-      x: (normalized.x - 0.5) * span * zoom + viewportSize.width / 2 + pan.width,
-      y: (normalized.y - 0.5) * span * zoom + viewportSize.height / 2 + pan.height
-    )
-  }
-
-  private static func priorityTier(
-    for placement: MemoryAtlasNodePlacement,
-    anchorNodeID: String?,
-    selectedNodeID: String?,
-    relatedNodeIDs: Set<String>,
-    matchingNodeIDs: Set<String>?
-  ) -> Int {
-    if placement.id == selectedNodeID {
-      return 0
-    } else if matchingNodeIDs?.contains(placement.id) == true {
-      return 1
-    } else if relatedNodeIDs.contains(placement.id) {
-      return 2
-    } else if placement.id == anchorNodeID {
-      return 3
-    } else {
-      return 4
-    }
-  }
-
-  /// Stable, allocation-light priority selection for gesture updates. The
-  /// layout already orders each cluster by salience, so a per-frame sort would
-  /// only spend main-thread time rediscovering that same order.
-  private static func priorityOrderedPrefix(
-    _ candidates: [MemoryAtlasNodePlacement],
-    limit: Int,
-    anchorNodeID: String?,
-    selectedNodeID: String?,
-    relatedNodeIDs: Set<String>,
-    matchingNodeIDs: Set<String>?,
-    includeBackgroundNodes: Bool
-  ) -> [MemoryAtlasNodePlacement] {
-    var tiers = Array(repeating: [MemoryAtlasNodePlacement](), count: 5)
-    for placement in candidates {
-      let tier = priorityTier(
-        for: placement,
-        anchorNodeID: anchorNodeID,
-        selectedNodeID: selectedNodeID,
-        relatedNodeIDs: relatedNodeIDs,
-        matchingNodeIDs: matchingNodeIDs
-      )
-      tiers[tier].append(placement)
-    }
-
-    var result: [MemoryAtlasNodePlacement] = []
-    result.reserveCapacity(min(limit, candidates.count))
-    let tierCount = includeBackgroundNodes ? tiers.count : 3
-    for tierIndex in 0..<tierCount where result.count < limit {
-      let remaining = limit - result.count
-      result.append(
-        contentsOf: fairPrefix(
-          tiers[tierIndex],
-          limit: remaining,
-          prioritizeCatalog: matchingNodeIDs != nil && tierIndex <= 1
-        )
-      )
-    }
-    return result
-  }
-
-  /// Labels admitted without collision, plus the subset drawn above their mark.
-  private struct AdmittedLabels {
-    var placements: [MemoryAtlasNodePlacement] = []
-    var aboveNodeIDs: Set<String> = []
-  }
-
-  private static func admitLabels(
-    _ candidates: [MemoryAtlasNodePlacement],
-    limit: Int,
-    viewportSize: CGSize,
-    zoom: CGFloat,
-    pan: CGSize,
-    compact: Bool,
-    forcedNodeIDs: Set<String>,
-    allowsFlipAbove: Bool
-  ) -> AdmittedLabels {
-    var result = AdmittedLabels()
-    var occupied: [CGRect] = []
-    result.placements.reserveCapacity(limit)
-    occupied.reserveCapacity(limit)
-
-    for placement in candidates {
-      let center = renderedPoint(
-        for: placement.normalizedPosition,
-        viewportSize: viewportSize,
-        zoom: zoom,
-        pan: pan
-      )
-      let estimatedWidth = min(
-        compact ? 112.0 : 152.0,
-        max(44.0, CGFloat(placement.node.label.count) * (compact ? 5.7 : 6.4) + 18)
-      )
-      let height: CGFloat = compact ? 22 : 26
-      let gap: CGFloat = compact ? 10 : 13
-      let x = center.x - estimatedWidth / 2
-      let below = CGRect(x: x, y: center.y + gap, width: estimatedWidth, height: height)
-        .insetBy(dx: -5, dy: -3)
-      // Flipping above is a second chance at the same name, not a second label.
-      // A dense atlas cannot afford one — there, a collision still means the
-      // name is dropped, because thousands of names never fit either way.
-      let above = CGRect(x: x, y: center.y - gap - height, width: estimatedWidth, height: height)
-        .insetBy(dx: -5, dy: -3)
-
-      let forced = forcedNodeIDs.contains(placement.id)
-      let fitsBelow = !occupied.contains { $0.intersects(below) }
-      let fitsAbove = allowsFlipAbove && !occupied.contains { $0.intersects(above) }
-
-      let chosen: CGRect
-      if fitsBelow || forced {
-        chosen = below
-      } else if fitsAbove {
-        chosen = above
-        result.aboveNodeIDs.insert(placement.id)
-      } else {
-        continue
-      }
-
-      result.placements.append(placement)
-      occupied.append(chosen)
-      if result.placements.count == limit { break }
-    }
-    return result
-  }
-}
-
-enum MemoryAtlasLayoutEngine {
-  static func makeSnapshot(
-    graph: KnowledgeGraphResponse,
-    userName: String?
-  ) -> MemoryAtlasSnapshot {
-    // Graph responses are external data. Coalesce duplicate identifiers at the
-    // boundary so a malformed server response cannot trap while building a UI.
-    var nodes = uniqueNodes(from: graph.atlasNodes)
-    let assertionNodeIDs = Set(nodes.map(\.id))
-    // A canonical memory without a verified relationship is still a real
-    // memory. It belongs in the atlas as a neutral, explicitly unconnected
-    // mark rather than being silently removed from the user's browser. Keep it
-    // out of the semantic layout below: using a type field or an inferred edge
-    // here would make an unsupported claim about what that memory relates to.
-    let catalogNodes = uniqueNodes(from: graph.catalogNodes ?? [])
-      .filter { !assertionNodeIDs.contains($0.id) }
-    let edges = uniqueEdges(from: graph.edges).filter {
-      assertionNodeIDs.contains($0.sourceId) && assertionNodeIDs.contains($0.targetId)
-    }
-    var degree: [String: Int] = [:]
-    for edge in edges {
-      degree[edge.sourceId, default: 0] += 1
-      degree[edge.targetId, default: 0] += 1
-    }
-
-    let owner = MemoryAtlasOwnerIdentity.resolve(nodes: nodes, userName: userName)
-    let anchor: KnowledgeGraphNode? = owner.anchor.map { rawAnchor in
-      guard let userName, !userName.isEmpty, MemoryAtlasOwnerIdentity.isSelfNode(rawAnchor)
-      else { return rawAnchor }
-      return KnowledgeGraphNode(
-        id: rawAnchor.id,
-        label: userName,
-        nodeType: rawAnchor.nodeType,
-        aliases: rawAnchor.aliases + [rawAnchor.label],
-        memoryIds: rawAnchor.memoryIds,
-        createdAt: rawAnchor.createdAt,
-        updatedAt: rawAnchor.updatedAt
-      )
-    }
-    if owner.isSynthetic, let anchor { nodes.append(anchor) }
-
-    guard !nodes.isEmpty || !catalogNodes.isEmpty else {
-      return MemoryAtlasSnapshot(nodes: [], edges: [], anchorNodeID: nil, clusterCenters: [:])
-    }
-
-    // Merge generic account-holder aliases into one identifiable center.
-    let collapsedIDs: Set<String> = {
-      guard let anchor else { return [] }
-      return Set(
-        nodes.filter { node in
-          guard node.id != anchor.id else { return false }
-          let label = node.label.lowercased()
-          let normalizedUserName = userName?.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-          if let normalizedUserName, !normalizedUserName.isEmpty, label == normalizedUserName {
-            return true
-          }
-          return MemoryAtlasOwnerIdentity.isSelfNode(node)
-        }.map(\.id)
-      )
-    }()
-
-    var grouped: [MemoryAtlasCluster: [KnowledgeGraphNode]] = [:]
-    for node in nodes where node.id != anchor?.id && !collapsedIDs.contains(node.id) {
-      grouped[cluster(for: node), default: []].append(node)
-    }
-    let activeClusters = MemoryAtlasCluster.allCases.filter { !(grouped[$0] ?? []).isEmpty }
-    let clusterCenters = MemoryAtlasCluster.centers(for: activeClusters)
-
-    let canonicalID: (String) -> String = { id in
-      guard let anchorID = anchor?.id, collapsedIDs.contains(id) else { return id }
-      return anchorID
-    }
-
-    // Every id that will actually receive a placement: the anchor plus every
-    // grouped node. An edge naming anything else is either a dangling server
-    // reference to a node absent from `graph.nodes`, or was already folded
-    // into the anchor by `canonicalID`.
-    var placeableNodeIDs: Set<String> = []
-    if let anchor { placeableNodeIDs.insert(anchor.id) }
-    for groupNodes in grouped.values {
-      for node in groupNodes { placeableNodeIDs.insert(node.id) }
-    }
-
-    // Reroute every edge onto its canonical (post-collapse) endpoints before
-    // merging parallel edges, so a relationship the collapsed self-node had
-    // with some entity is considered for merging against the anchor's own
-    // edge to that same entity rather than kept as a separate pair.
-    let reroutedEdges: [KnowledgeGraphEdge] = edges.compactMap { edge in
-      let sourceId = canonicalID(edge.sourceId)
-      let targetId = canonicalID(edge.targetId)
-      // A relationship the collapsed self-node had with the anchor becomes a
-      // self-loop after rerouting — drop it rather than draw a node to itself.
-      guard sourceId != targetId else { return nil }
-      guard placeableNodeIDs.contains(sourceId), placeableNodeIDs.contains(targetId) else { return nil }
-      // Only rebuild the edge when an endpoint actually moved, so neighbor and
-      // evidence lookups resolve against the single center node.
-      if sourceId == edge.sourceId && targetId == edge.targetId { return edge }
-      return KnowledgeGraphEdge(
-        id: edge.id,
-        sourceId: sourceId,
-        targetId: targetId,
-        label: edge.label,
-        memoryIds: edge.memoryIds,
-        createdAt: edge.createdAt
-      )
-    }
-
-    // The server can emit several edges for what is really one relationship
-    // between two entities — e.g. "includes" and "with" for the same pair —
-    // which otherwise doubles both the drawn line and the inspector's
-    // Connections row for it. Merge by unordered endpoint pair before laying
-    // out nodes, so a node's distinct-neighbor count (not its raw edge-touch
-    // count) is what decides leaf/isolate placement below.
-    let mergedEdges = mergeParallelEdges(reroutedEdges)
-
-    var neighborIDs: [String: Set<String>] = [:]
-    for merged in mergedEdges {
-      neighborIDs[merged.edge.sourceId, default: []].insert(merged.edge.targetId)
-      neighborIDs[merged.edge.targetId, default: []].insert(merged.edge.sourceId)
-    }
-
-    // Relatedness, from both signals the client already holds: the edges the
-    // extractor drew, and the memories two entities were extracted from
-    // together. The second matters more than it sounds — it finds
-    // relationships nobody wrote down, which is most of them.
-    //
-    // The anchor is deliberately absent: they appear in nearly every memory, so
-    // projecting them would make "was in a memory with you" the strongest
-    // signal about every entity, which is the one thing that is true of all of
-    // them. `coOccurrenceLinks` enforces this itself; not building the entry
-    // just avoids collecting memories only to drop them.
-    var memoryIDsByNodeID: [String: [String]] = [:]
-    for groupNodes in grouped.values {
-      for node in groupNodes { memoryIDsByNodeID[node.id] = node.memoryIds }
-    }
-
-    let relatedness =
-      MemoryAtlasForceLayout.explicitLinks(
-        edges: mergedEdges.map {
-          (
-            sourceID: $0.edge.sourceId, targetID: $0.edge.targetId,
-            memoryCount: $0.edge.memoryIds.count
-          )
-        })
-      + MemoryAtlasForceLayout.coOccurrenceLinks(
-        memoryIDsByNodeID: memoryIDsByNodeID, excluding: anchor?.id)
-
-    // Type stops deciding where a node goes and becomes a weak field it can
-    // overrule: the constellation centres are now only somewhere a node drifts
-    // toward when nothing it is related to pulls harder.
-    var typeTargets: [String: CGPoint] = [:]
-    for cluster in activeClusters {
-      guard let center = clusterCenters[cluster] else { continue }
-      for node in grouped[cluster] ?? [] { typeTargets[node.id] = center }
-    }
-
-    let layout = MemoryAtlasForceLayout.layout(
-      nodeIDs: Array(placeableNodeIDs),
-      links: relatedness,
-      anchorID: anchor?.id,
-      typeTargets: typeTargets,
-      area: Self.layoutArea)
-
-    var placements: [MemoryAtlasNodePlacement] = []
-    if let anchor {
-      placements.append(
-        MemoryAtlasNodePlacement(
-          node: anchor,
-          cluster: nil,
-          normalizedPosition: layout.positions[anchor.id] ?? MemoryAtlasCluster.starCenter,
-          degree: neighborIDs[anchor.id]?.count ?? 0,
-          clusterRank: 0,
-          isCatalog: false
-        )
-      )
-    }
-
-    // Salience order no longer decides position, but it still decides which
-    // names win a crowded viewport, so the rank the render planner reads is
-    // computed exactly as before.
-    for cluster in activeClusters {
-      let sorted = (grouped[cluster] ?? []).sorted {
-        let lhsScore = salience(node: $0, degree: degree[$0.id] ?? 0)
-        let rhsScore = salience(node: $1, degree: degree[$1.id] ?? 0)
-        if lhsScore == rhsScore { return $0.label.localizedCaseInsensitiveCompare($1.label) == .orderedAscending }
-        return lhsScore > rhsScore
-      }
-
-      for (index, node) in sorted.enumerated() {
-        placements.append(
-          MemoryAtlasNodePlacement(
-            node: node,
-            cluster: cluster,
-            normalizedPosition: layout.positions[node.id] ?? MemoryAtlasCluster.starCenter,
-            degree: neighborIDs[node.id]?.count ?? 0,
-            clusterRank: index,
-            isCatalog: false
-          )
-        )
-      }
-    }
-    let assertionPositions = Dictionary(lastWriteWins: placements.map { ($0.id, $0.normalizedPosition) })
-    let catalogPositions = MemoryAtlasCatalogLayout.positions(
-      catalog: catalogNodes,
-      assertions: nodes.filter { $0.id != anchor?.id && !collapsedIDs.contains($0.id) },
-      communities: layout.communities,
-      assertionPositions: assertionPositions,
-      ownerID: anchor?.id,
-      area: Self.layoutArea)
-    for (index, node) in catalogNodes.sorted(by: { $0.id < $1.id }).enumerated() {
-      placements.append(
-        MemoryAtlasNodePlacement(
-          node: node,
-          cluster: nil,
-          normalizedPosition: catalogPositions[node.id] ?? MemoryAtlasCluster.starCenter,
-          degree: 0,
-          clusterRank: index + 1,
-          isCatalog: true
-        )
-      )
-    }
-    let positions = Dictionary(lastWriteWins: placements.map { ($0.id, $0.normalizedPosition) })
-    let clusters = Dictionary(
-      lastWriteWins: placements.compactMap { placement in
-        placement.cluster.map { (placement.id, $0) }
-      })
-
-    let edgePlacements = mergedEdges.map { merged -> MemoryAtlasEdgePlacement in
-      let source = positions[merged.edge.sourceId] ?? MemoryAtlasCluster.starCenter
-      let target = positions[merged.edge.targetId] ?? MemoryAtlasCluster.starCenter
-      let cluster = clusters[merged.edge.sourceId] ?? clusters[merged.edge.targetId] ?? .concept
-      let sourceGroup = layout.communities[merged.edge.sourceId]
-      return MemoryAtlasEdgePlacement(
-        edge: merged.edge,
-        source: source,
-        target: target,
-        cluster: cluster,
-        relationshipLabels: merged.relationshipLabels,
-        weight: merged.weight,
-        withinNeighbourhood: sourceGroup != nil
-          && sourceGroup == layout.communities[merged.edge.targetId]
-      )
-    }
-
-    return MemoryAtlasSnapshot(
-      nodes: placements,
-      edges: edgePlacements,
-      anchorNodeID: anchor?.id,
-      clusterCenters: centroids(of: placements, activeClusters: activeClusters),
-      neighbourhoods: neighbourhoods(communities: layout.communities, placements: placements)
-    )
-  }
-
-  /// Turns the layout's group numbers into regions the map can name.
-  ///
-  /// Two rules keep this from labelling noise. A region must be big enough to
-  /// be a place — a modularity pass on a real account emits a long tail of
-  /// three-entity groups, and captioning those would bury the handful that
-  /// matter. And its extent is the distance covering most of its members, not
-  /// all of them, so one entity that drifted across the map does not inflate a
-  /// tight neighbourhood into a claim over half the canvas.
-  static func neighbourhoods(
-    communities: [String: Int],
-    placements: [MemoryAtlasNodePlacement],
-    captionMembers: Int = captionNames
-  ) -> [MemoryAtlasNeighbourhood] {
-    guard !communities.isEmpty else { return [] }
-    let placementByID = Dictionary(lastWriteWins: placements.map { ($0.id, $0) })
-
-    var membersByGroup: [Int: [String]] = [:]
-    for id in communities.keys.sorted() {
-      guard let group = communities[id], placementByID[id] != nil else { continue }
-      membersByGroup[group, default: []].append(id)
-    }
-
-    // Scaled to the map: six entities is a neighbourhood on a small account and
-    // rounding error on a large one.
-    let floor = max(6, Int(Double(placementByID.count) * 0.015))
-
-    // Every group over the floor competes for ground, including the ones too
-    // small or too crowded to end up captioned. Leaving one out would not make
-    // it disappear — it would hand its land to whichever neighbour is nearest,
-    // and draw a coast through the middle of a group that is still on screen.
-    var contenders: [Int: [CGPoint]] = [:]
-    for group in membersByGroup.keys.sorted() {
-      let members = membersByGroup[group] ?? []
-      guard members.count >= floor else { continue }
-      contenders[group] = members.compactMap { placementByID[$0]?.normalizedPosition }
-    }
-    let coastlines = MemoryAtlasIslands.coastlines(members: contenders)
-
-    var regions: [MemoryAtlasNeighbourhood] = []
-    for group in membersByGroup.keys.sorted() {
-      let members = membersByGroup[group] ?? []
-      guard members.count >= floor else { continue }
-      let points = members.compactMap { placementByID[$0]?.normalizedPosition }
-      guard !points.isEmpty else { continue }
-      // A group whose members are spread too thin to hold any ground has no
-      // territory to draw and no border to claim.
-      let coastline = coastlines[group] ?? []
-      guard !coastline.isEmpty else { continue }
-
-      let center = CGPoint(
-        x: points.map(\.x).reduce(0, +) / CGFloat(points.count),
-        y: points.map(\.y).reduce(0, +) / CGFloat(points.count)
-      )
-      let distances = points.map { hypot($0.x - center.x, $0.y - center.y) }.sorted()
-      let radius = distances[min(distances.count - 1, Int(Double(distances.count) * 0.8))]
-
-      let own = Set(members)
-      var inside = 0
-      var mine = 0
-      for placement in placements
-      where memoryAtlasCoastlineContains(coastline, placement.normalizedPosition) {
-        inside += 1
-        if own.contains(placement.id) { mine += 1 }
-      }
-
-      // Most connected first, so the caption names the entities a person would
-      // actually recognise the region by.
-      let ranked =
-        members
-        .compactMap { placementByID[$0] }
-        .sorted {
-          if $0.degree == $1.degree {
-            return $0.node.label.localizedCaseInsensitiveCompare($1.node.label) == .orderedAscending
-          }
-          return $0.degree > $1.degree
-        }
-
-      regions.append(
-        MemoryAtlasNeighbourhood(
-          id: group,
-          memberIDs: members,
-          caption: caption(from: ranked, count: captionMembers),
-          center: center,
-          radius: max(radius, 0.01),
-          coastline: coastline,
-          purity: inside == 0 ? 0 : Double(mine) / Double(inside)
-        )
-      )
-    }
-
-    return regions.sorted {
-      $0.memberCount == $1.memberCount ? $0.id < $1.id : $0.memberCount > $1.memberCount
-    }
-  }
-
-  /// The longest a member's label may be and still work as part of a region's
-  /// name. Roughly a long proper noun — "Ho Chi Minh City", "Omi Desktop App".
-  static let captionNameCeiling = 26
-
-  /// The most a region's name may be and still be read at a glance.
-  ///
-  /// A place on a map is one word you recognise, not a list. Three names joined
-  /// by separators ran to the width of a paragraph, truncated mid-word into
-  /// things like "ORACLE ACCESS · CONSULT-ORACLE · DAVI…", and took longer to
-  /// parse than the dots underneath it — so the map read as a wall of log lines
-  /// laid over a graph. One name is a landmark: you either know it or you zoom
-  /// in, and both are faster than reading three.
-  static let captionNames = 1
-
-  /// Picks the few names that describe a region.
-  ///
-  /// Connectedness alone is not enough, because the extractor does not only
-  /// mint names. It also mints entities whose label is an entire sentence — a
-  /// calendar invite's full subject line, a GitHub issue title with its quotes
-  /// intact — and those are frequently a region's best-connected members. Taken
-  /// literally, one of them produced a region caption 1,200 points wide that
-  /// covered a quarter of the map and named nothing.
-  ///
-  /// So the caption prefers members whose labels read as names, and only falls
-  /// back to truncating a sentence when a region has nothing shorter to offer.
-  /// Reading like a name means short *and* few words: "Omi Desktop App" is a
-  /// landmark and "explicit request by David" is a sentence that happens to fit.
-  static func caption(
-    from ranked: [MemoryAtlasNodePlacement], count: Int, ceiling: Int = captionNameCeiling
-  ) -> String {
-    func words(_ label: String) -> Int {
-      label.split(whereSeparator: \.isWhitespace).count
-    }
-    var chosen =
-      ranked
-      .filter { $0.node.label.count <= ceiling && words($0.node.label) <= 3 }
-      .prefix(count).map(\.node.label)
-    if chosen.count < count {
-      let already = Set(chosen)
-      for placement in ranked
-      where chosen.count < count && !already.contains(placement.node.label)
-        && placement.node.label.count <= ceiling
-      {
-        chosen.append(placement.node.label)
-      }
-    }
-    if chosen.count < count {
-      let already = Set(chosen)
-      for placement in ranked where chosen.count < count {
-        let label = placement.node.label
-        guard !already.contains(label), label.count > ceiling else { continue }
-        chosen.append(String(label.prefix(ceiling - 1)).trimmingCharacters(in: .whitespaces) + "…")
-      }
-    }
-    return chosen.joined(separator: " · ")
-  }
-
-  /// The region the relaxed map may occupy. Isolates are parked in the margin
-  /// outside it, so it stops short of the canvas edge.
-  ///
-  /// Square, because the canvas it lands on is projected square. The map uses
-  /// almost the available height: a smaller square created a visible box of
-  /// empty space around dense accounts even though the underlying layout was
-  /// circular. A wide box would stretch that circle into a desktop smear.
-  static let layoutArea = CGRect(x: 0.06, y: 0.06, width: 0.88, height: 0.88)
-
-  /// How many points one unit of normalized map space is worth on screen.
-  ///
-  /// Deliberately one number rather than one per axis. The relaxation works in
-  /// a square and knows nothing about the window; projecting each axis onto its
-  /// own side of the viewport distorted whatever shape it found by the window's
-  /// aspect ratio, which on a wide desktop meant every account was drawn as a
-  /// horizontal band. Taking the shorter side for both axes leaves the map's
-  /// proportions alone and spends the extra width as margin — which is what the
-  /// graph views this one is measured against do.
-  static func projectionSpan(of size: CGSize) -> CGFloat { min(size.width, size.height) }
-
-  /// One phrasing for "how big is this map", so the header and the timeline
-  /// footer cannot describe the same thing in two different ways.
-  static func countLabel(entities: Int, memories: Int? = nil, connections: Int) -> String {
-    let entityLabel = "\(entities) entit\(entities == 1 ? "y" : "ies")"
-    let connectionLabel = "\(connections) connection\(connections == 1 ? "" : "s")"
-    guard let memories else { return "\(entityLabel) · \(connectionLabel)" }
-    return "\(entityLabel) · \(memories) memor\(memories == 1 ? "y" : "ies") · \(connectionLabel)"
-  }
-
-  /// Whether the account holder's own connections should recede into the
-  /// background rather than being drawn like every other relationship.
-  ///
-  /// They are the least informative lines on the map — everything is connected
-  /// to you — and by far the most numerous, so at full strength a few hundred
-  /// straight spokes cross every neighbourhood and the map reads as a star no
-  /// matter where its entities actually sit. The exception is the one moment
-  /// they *are* the subject: selecting the account holder is asking "what am I
-  /// connected to", and the answer has to be drawn.
-  static func anchorConnectionsRecede(anchorID: String?, selectedNodeID: String?) -> Bool {
-    guard let anchorID else { return false }
-    return selectedNodeID != anchorID
-  }
-
-  /// Where each type actually ended up, rather than where it was assigned.
-  ///
-  /// A type no longer owns a region of the canvas, so labelling a fixed petal
-  /// "People" would point at empty space. The caption instead sits at the mean
-  /// position of that type's entities — which is honest, and lands somewhere
-  /// useful precisely because the weak type field keeps each type loosely
-  /// gathered.
-  static func centroids(
-    of placements: [MemoryAtlasNodePlacement],
-    activeClusters: [MemoryAtlasCluster]
-  ) -> [MemoryAtlasCluster: CGPoint] {
-    var totals: [MemoryAtlasCluster: (x: CGFloat, y: CGFloat, count: Int)] = [:]
-    for placement in placements {
-      guard let cluster = placement.cluster else { continue }
-      var running = totals[cluster] ?? (0, 0, 0)
-      running.x += placement.normalizedPosition.x
-      running.y += placement.normalizedPosition.y
-      running.count += 1
-      totals[cluster] = running
-    }
-
-    var result: [MemoryAtlasCluster: CGPoint] = [:]
-    for cluster in activeClusters {
-      guard let running = totals[cluster], running.count > 0 else { continue }
-      result[cluster] = CGPoint(
-        x: running.x / CGFloat(running.count),
-        y: running.y / CGFloat(running.count))
-    }
-    return result
-  }
-
-  /// One placement's worth of collapsed parallel edges: the representative
-  /// edge (first-seen orientation and id), every distinct relationship label
-  /// merged in, the union of citing memory ids, and how many raw edges fed
-  /// into it.
-  private struct MergedEdge {
-    private(set) var edge: KnowledgeGraphEdge
-    private(set) var relationshipLabels: [String]
-    private(set) var weight: Int
-    private var seenNormalizedLabels: Set<String>
-    private var memoryIds: [String]
-    private var seenMemoryIds: Set<String>
-
-    init(edge: KnowledgeGraphEdge) {
-      self.edge = edge
-      relationshipLabels = [edge.label]
-      seenNormalizedLabels = [MemoryAtlasLayoutEngine.normalizeRelationship(edge.label)]
-      weight = 1
-      memoryIds = edge.memoryIds
-      seenMemoryIds = Set(edge.memoryIds)
-    }
-
-    mutating func merge(_ other: KnowledgeGraphEdge) {
-      weight += 1
-      let normalized = MemoryAtlasLayoutEngine.normalizeRelationship(other.label)
-      if seenNormalizedLabels.insert(normalized).inserted {
-        relationshipLabels.append(other.label)
-      }
-      var memoryIdsChanged = false
-      for id in other.memoryIds where seenMemoryIds.insert(id).inserted {
-        memoryIds.append(id)
-        memoryIdsChanged = true
-      }
-      guard memoryIdsChanged else { return }
-      // Keep the representative edge's own `memoryIds` in sync with the
-      // union so a caller reading `edge.memoryIds` directly (evidence
-      // resolution) sees every citing memory, not just the first edge's.
-      edge = KnowledgeGraphEdge(
-        id: edge.id,
-        sourceId: edge.sourceId,
-        targetId: edge.targetId,
-        label: edge.label,
-        memoryIds: memoryIds,
-        createdAt: edge.createdAt
-      )
-    }
-  }
-
-  /// Collapses edges that share the same unordered endpoint pair into one
-  /// `MergedEdge`, preserving the order pairs were first encountered so the
-  /// result is deterministic for a fixed input order.
-  private static func mergeParallelEdges(_ edges: [KnowledgeGraphEdge]) -> [MergedEdge] {
-    var order: [String] = []
-    var groups: [String: MergedEdge] = [:]
-    for edge in edges {
-      let pairKey =
-        edge.sourceId <= edge.targetId
-        ? "\(edge.sourceId)\u{0}\(edge.targetId)"
-        : "\(edge.targetId)\u{0}\(edge.sourceId)"
-      if var existing = groups[pairKey] {
-        existing.merge(edge)
-        groups[pairKey] = existing
-      } else {
-        groups[pairKey] = MergedEdge(edge: edge)
-        order.append(pairKey)
-      }
-    }
-    return order.compactMap { groups[$0] }
-  }
-
-  private static func uniqueNodes(from nodes: [KnowledgeGraphNode]) -> [KnowledgeGraphNode] {
-    var seenIDs: Set<String> = []
-    let newestFirst = nodes.reversed().compactMap { node -> KnowledgeGraphNode? in
-      seenIDs.insert(node.id).inserted ? node : nil
-    }
-    return Array(newestFirst.reversed())
-  }
-
-  private static func uniqueEdges(from edges: [KnowledgeGraphEdge]) -> [KnowledgeGraphEdge] {
-    var seenIDs: Set<String> = []
-    let newestFirst = edges.reversed().compactMap { edge -> KnowledgeGraphEdge? in
-      seenIDs.insert(edge.id).inserted ? edge : nil
-    }
-    return Array(newestFirst.reversed())
-  }
-
-  static func cluster(for node: KnowledgeGraphNode) -> MemoryAtlasCluster {
-    switch node.nodeType {
-    case .person: return .person
-    case .organization: return .organization
-    case .place: return .place
-    case .thing: return .thing
-    case .concept: return .concept
-    }
-  }
-
-  static func relationshipDisplayName(_ rawValue: String) -> String {
-    normalizeRelationship(rawValue).replacingOccurrences(of: "_", with: " ")
-  }
-
-  /// Display text for a placement whose parallel edges carry more than one
-  /// distinct verb. A merged pair reads as one relationship described two
-  /// ways ("includes & with"), not two separate relationships, so the
-  /// Connections list shows a single combined row per neighbor rather than
-  /// one row per verb — that keeps the list's row count matching the actual
-  /// number of distinct entities a node touches.
-  static func combinedRelationshipDisplayName(_ rawValues: [String]) -> String {
-    rawValues.map(relationshipDisplayName).joined(separator: " & ")
-  }
-
-  private static func normalizeRelationship(_ value: String) -> String {
-    value
-      .trimmingCharacters(in: .whitespacesAndNewlines)
-      .lowercased()
-      .replacingOccurrences(of: "-", with: "_")
-      .replacingOccurrences(of: " ", with: "_")
-  }
-
-  private static func salience(node: KnowledgeGraphNode, degree: Int) -> Int {
-    let typeBonus: Int
-    switch node.nodeType {
-    case .organization: typeBonus = 16
-    case .person: typeBonus = 10
-    case .concept: typeBonus = 4
-    case .place: typeBonus = 3
-    case .thing: typeBonus = 0
-    }
-    let genericPenalty = genericLabels.contains(node.label.lowercased()) ? 10_000 : 0
-    return degree * 20 + node.memoryIds.count * 4 + min(node.aliases.count, 3) + typeBonus - genericPenalty
-  }
-
-  private static let genericLabels: Set<String> = [
-    "app", "apps", "user", "calendar event", "document", "documents", "download", "downloads",
-  ]
-
-}
-
-// MARK: - Node Inspector
-
-/// One memory backing a selected entity, flattened so the atlas surface does
-/// not depend on the memories layer. The panel renders these directly instead
-/// of navigating away, which is what makes the Brain Map a place you can stay
-/// in while reading what an entity is made of.
-struct MemoryAtlasEvidence: Identifiable, Equatable {
-  let id: String
-  let content: String
-  let createdAt: Date?
-
-  /// Resolves cited memory ids against whatever the memories layer has loaded,
-  /// preserving that order so the inspector reads newest-first like the list.
-  /// Ids with no loaded memory are simply absent — the panel reports the gap
-  /// rather than the surface pretending the entity has less evidence than it
-  /// does.
-  static func resolve(_ ids: [String], in memories: [ServerMemory]) -> [MemoryAtlasEvidence] {
-    let wanted = Set(ids)
-    guard !wanted.isEmpty else { return [] }
-    return memories.filter { wanted.contains($0.id) }.map {
-      MemoryAtlasEvidence(id: $0.id, content: $0.content, createdAt: $0.createdAt)
-    }
-  }
-}
-
-/// What the inspector is describing.
-///
-/// The Brain Map has two things worth clicking — the entities and the
-/// connections between them — and both answer the same three questions: what
-/// is this, what does it touch, and which memories produced it. One panel
-/// shape for both keeps traversal continuous instead of making a relationship
-/// a dead end.
-enum MemoryAtlasInspectorSubject: Equatable {
-  case entity(title: String, typeName: String?, connectionSummary: String)
-  case relationship(sourceLabel: String, targetLabel: String, verb: String)
-
-  var title: String {
-    switch self {
-    case .entity(let title, _, _):
-      return title
-    case .relationship(let sourceLabel, let targetLabel, _):
-      return "\(sourceLabel) → \(targetLabel)"
-    }
-  }
-
-  var subtitle: String {
-    switch self {
-    case .entity(_, let typeName, let connectionSummary):
-      return [typeName, connectionSummary].compactMap { $0 }.joined(separator: " · ")
-    case .relationship(_, _, let verb):
-      return verb
-    }
-  }
-
-  /// Header for the section listing what this subject connects to.
-  var relatedSectionTitle: String {
-    switch self {
-    case .entity: return "Connections"
-    case .relationship: return "Between"
-    }
-  }
-}
-
-/// Right-hand inspector for the current selection.
-///
-/// Replaces the "View evidence" jump to Memories: selecting a node used to
-/// change pages, which lost the camera, the time cursor, and the selection.
-struct MemoryAtlasDetailPanel: View {
-  let subject: MemoryAtlasInspectorSubject
-  let accent: Color
-  let related: [MemoryAtlasRelationshipRow]
-  let evidence: [MemoryAtlasEvidence]
-  let evidenceIsLoading: Bool
-  let unresolvedEvidenceCount: Int
-  let onOpenRelated: (MemoryAtlasRelationshipRow) -> Void
-  /// Opens a cited memory on the Memories page with its detail panel showing.
-  /// The inspector deliberately shows the whole memory in place, but a memory
-  /// is also a thing you act on — edit it, check its provenance, delete it —
-  /// and none of that belongs in a graph inspector.
-  let onOpenMemory: (MemoryAtlasEvidence) -> Void
-  /// Present once the user has followed a connection. Traversal without a way
-  /// back turns every hop into a dead end you can only escape by hunting for
-  /// the previous dot on the canvas.
-  let onBack: (() -> Void)?
-  let onFocus: () -> Void
-  let onClose: () -> Void
-
-  @State private var hoveredRelatedID: String?
-  @State private var hoveredEvidenceID: String?
-
-  var body: some View {
-    VStack(alignment: .leading, spacing: 0) {
-      header
-
-      Divider().overlay(Ink.separator.opacity(0.2))
-
-      ScrollView {
-        VStack(alignment: .leading, spacing: 18) {
-          if !related.isEmpty {
-            section(subject.relatedSectionTitle) {
-              VStack(alignment: .leading, spacing: 2) {
-                ForEach(related) { row in
-                  relatedRow(row)
-                }
-              }
-              // Rows carry their own inset so the hover fill reads as a
-              // target; pulling it back keeps their text on the same margin
-              // as the section title.
-              .padding(.horizontal, -6)
-            }
-          }
-
-          section("From your memories") {
-            evidenceSection
-          }
-        }
-        .padding(16)
-      }
-    }
-    .frame(width: 320)
-    .background(Ink.rowFill)
-    .overlay(alignment: .leading) {
-      Rectangle().fill(Ink.separator.opacity(0.25)).frame(width: 1)
-    }
-    .accessibilityIdentifier("memory_atlas_detail_panel")
-  }
-
-  @ViewBuilder
-  private var evidenceSection: some View {
-    if evidenceIsLoading && evidence.isEmpty {
-      HStack(spacing: 7) {
-        ProgressView()
-          .controlSize(.small)
-          .tint(Ink.secondary)
-        Text("Reading your memories…")
-          .scaledFont(size: 11)
-          .foregroundColor(Ink.secondary)
-      }
-    } else if evidence.isEmpty && unresolvedEvidenceCount == 0 {
-      Text("Source memories are still being linked for this entity.")
-        .scaledFont(size: 11)
-        .foregroundColor(Ink.secondary)
-        .fixedSize(horizontal: false, vertical: true)
-    } else {
-      ForEach(evidence) { item in
-        evidenceRow(item)
-      }
-      if unresolvedEvidenceCount > 0 {
-        // Evidence resolves against the full local cache, so a remaining gap
-        // means the cited memory genuinely is not on this device — not that
-        // the list simply had not paged far enough yet.
-        Text(
-          "\(unresolvedEvidenceCount) cited memor\(unresolvedEvidenceCount == 1 ? "y" : "ies") could not be found on this device"
-        )
-        .scaledFont(size: 10)
-        .foregroundColor(Ink.secondary)
-        .fixedSize(horizontal: false, vertical: true)
-      }
-    }
-  }
-
-  private var header: some View {
-    HStack(alignment: .top, spacing: 10) {
-      if let onBack {
-        Button(action: onBack) {
-          Image(systemName: "chevron.left")
-            .scaledFont(size: 11, weight: .semibold)
-            .foregroundColor(Ink.secondary)
-            .frame(width: 22, height: 30)
-            .contentShape(Rectangle())
-        }
-        .buttonStyle(.plain)
-        .help("Back to the previous entity")
-        .accessibilityIdentifier("memory_atlas_inspector_back")
-      }
-
-      Circle()
-        .fill(accent.opacity(0.14))
-        .overlay(Circle().stroke(accent, lineWidth: 1.5))
-        .frame(width: 30, height: 30)
-
-      VStack(alignment: .leading, spacing: 3) {
-        Text(subject.title)
-          .scaledFont(size: 15, weight: .semibold)
-          .foregroundColor(Ink.primary)
-          .fixedSize(horizontal: false, vertical: true)
-        Text(subject.subtitle)
-          .scaledFont(size: 11)
-          .foregroundColor(Ink.secondary)
-          .fixedSize(horizontal: false, vertical: true)
-      }
-
-      Spacer(minLength: 4)
-
-      Button(action: onFocus) {
-        Image(systemName: "scope")
-          .scaledFont(size: 11, weight: .medium)
-          .foregroundColor(Ink.secondary)
-          .frame(width: 24, height: 24)
-          .contentShape(Rectangle())
-      }
-      .buttonStyle(.plain)
-      .help("Center the map on this entity")
-      .accessibilityIdentifier("memory_atlas_focus_selection")
-
-      Button(action: onClose) {
-        Image(systemName: "xmark")
-          .scaledFont(size: 10, weight: .semibold)
-          .foregroundColor(Ink.secondary)
-          .frame(width: 24, height: 24)
-          .contentShape(Rectangle())
-      }
-      .buttonStyle(.plain)
-      .help("Clear selection (Esc)")
-      .accessibilityIdentifier("memory_atlas_clear_selection")
-    }
-    .padding(.horizontal, 16)
-    .padding(.vertical, 14)
-  }
-
-  @ViewBuilder
-  private func section<Content: View>(
-    _ title: String,
-    @ViewBuilder content: () -> Content
-  ) -> some View {
-    VStack(alignment: .leading, spacing: 8) {
-      Text(title.uppercased())
-        .scaledFont(size: 9.5, weight: .semibold)
-        .foregroundColor(Ink.secondary)
-        .tracking(0.6)
-      content()
-    }
-    .frame(maxWidth: .infinity, alignment: .leading)
-  }
-
-  /// A connection is a way through the graph, not a caption. Clicking one
-  /// moves the inspector to the entity on the other end, so following a chain
-  /// of relationships never requires hunting for the next dot on the canvas.
-  private func relatedRow(_ row: MemoryAtlasRelationshipRow) -> some View {
-    let isHovered = hoveredRelatedID == row.id
-    return Button {
-      onOpenRelated(row)
-    } label: {
-      HStack(alignment: .top, spacing: 8) {
-        Circle()
-          .fill(row.accent)
-          .frame(width: 5, height: 5)
-          .padding(.top, 5)
-        VStack(alignment: .leading, spacing: 1) {
-          Text(row.otherLabel)
-            .scaledFont(size: 12, weight: .medium)
-            .foregroundColor(isHovered ? Ink.primary : Ink.secondary)
-            .fixedSize(horizontal: false, vertical: true)
-          Text(row.relationship)
-            .scaledFont(size: 10)
-            .foregroundColor(Ink.secondary)
-            .fixedSize(horizontal: false, vertical: true)
-        }
-        Spacer(minLength: 4)
-        Image(systemName: "chevron.right")
-          .scaledFont(size: 9, weight: .semibold)
-          .foregroundColor(Ink.secondary)
-          .opacity(isHovered ? 1 : 0)
-          .padding(.top, 3)
-      }
-      .padding(.vertical, 5)
-      .padding(.horizontal, 6)
-      .frame(maxWidth: .infinity, alignment: .leading)
-      .background(
-        RoundedRectangle(cornerRadius: 7, style: .continuous)
-          .fill(isHovered ? Ink.rowFill.opacity(0.7) : Color.clear)
-      )
-      .contentShape(Rectangle())
-    }
-    .buttonStyle(.plain)
-    .onHover { hovering in
-      if hovering {
-        hoveredRelatedID = row.id
-      } else if hoveredRelatedID == row.id {
-        hoveredRelatedID = nil
-      }
-    }
-    .help("Open \(row.otherLabel)")
-    .accessibilityIdentifier("memory_atlas_connection_row")
-  }
-
-  private func evidenceRow(_ item: MemoryAtlasEvidence) -> some View {
-    let isHovered = hoveredEvidenceID == item.id
-    return Button {
-      onOpenMemory(item)
-    } label: {
-      VStack(alignment: .leading, spacing: 4) {
-        Text(item.content)
-          .scaledFont(size: 11.5)
-          .foregroundColor(isHovered ? Ink.primary : Ink.secondary)
-          .fixedSize(horizontal: false, vertical: true)
-          .frame(maxWidth: .infinity, alignment: .leading)
-        // The date holds this row whether or not the cursor is here, so the
-        // "Open" affordance can fade in without the row changing height.
-        HStack(spacing: 4) {
-          if let createdAt = item.createdAt {
-            Text(createdAt.formatted(date: .abbreviated, time: .shortened))
-              .scaledFont(size: 9.5)
-              .foregroundColor(Ink.secondary)
-          }
-          Spacer(minLength: 4)
-          HStack(spacing: 3) {
-            Text("Open")
-              .scaledFont(size: 9.5, weight: .medium)
-            Image(systemName: "arrow.up.right")
-              .scaledFont(size: 8, weight: .semibold)
-          }
-          .foregroundColor(Ink.secondary)
-          .opacity(isHovered ? 1 : 0)
-        }
-      }
-      .frame(maxWidth: .infinity, alignment: .leading)
-      .padding(10)
-      .background(
-        RoundedRectangle(cornerRadius: 10, style: .continuous)
-          .fill(Ink.rowFill.opacity(isHovered ? 0.95 : 0.6))
-      )
-      .contentShape(Rectangle())
-    }
-    .buttonStyle(.plain)
-    .onHover { hovering in
-      if hovering {
-        hoveredEvidenceID = item.id
-      } else if hoveredEvidenceID == item.id {
-        hoveredEvidenceID = nil
-      }
-    }
-    .help("Open this memory on the Memories page")
-    .accessibilityIdentifier("memory_atlas_evidence_row")
-  }
-}
-
-struct MemoryAtlasRelationshipRow: Identifiable, Equatable {
-  let id: String
-  /// Where clicking this row goes. Without it the inspector could name the
-  /// entity on the other end but not open it.
-  let otherNodeID: String
-  let otherLabel: String
-  let relationship: String
-  let accent: Color
-}
-
-/// Picking a painted connection out of the atlas.
-///
-/// Lives outside the view so the two rules that matter — a line is only picked
-/// when the click is genuinely on it, and the nearest one wins — are testable
-/// without a window or a gesture.
-enum MemoryAtlasHitTesting {
-  struct Segment {
-    let id: String
-    let start: CGPoint
-    let end: CGPoint
-  }
-
-  static func distance(from point: CGPoint, toSegmentFrom start: CGPoint, to end: CGPoint) -> CGFloat {
-    let run = end.x - start.x
-    let rise = end.y - start.y
-    let lengthSquared = run * run + rise * rise
-    // A zero-length segment is a point; projecting onto it would divide by zero.
-    guard lengthSquared > 0 else { return hypot(point.x - start.x, point.y - start.y) }
-    let projection = ((point.x - start.x) * run + (point.y - start.y) * rise) / lengthSquared
-    let clamped = min(1, max(0, projection))
-    let nearest = CGPoint(x: start.x + clamped * run, y: start.y + clamped * rise)
-    return hypot(point.x - nearest.x, point.y - nearest.y)
-  }
-
-  /// How far off a painted line a click may land and still count, in points.
-  /// Connections are drawn under 2pt wide, so this is forgiveness for aim, not
-  /// a hit area wide enough to steal clicks from a neighbouring line.
-  static let connectionTolerance: CGFloat = 5
-
-  static func nearestSegment(
-    to point: CGPoint,
-    among segments: [Segment],
-    within tolerance: CGFloat
-  ) -> String? {
-    var best: (id: String, distance: CGFloat)?
-    for segment in segments {
-      let distance = distance(from: point, toSegmentFrom: segment.start, to: segment.end)
-      guard distance <= tolerance else { continue }
-      if best.map({ distance < $0.distance }) ?? true {
-        best = (segment.id, distance)
-      }
-    }
-    return best?.id
-  }
-}
-
 // MARK: - Canonical Atlas Containers
 
 /// Holds the canvas until a complete graph is ready, so the first visit cannot
