+++ desktop/macos/changelog/unreleased/20260901-memory-atlas-architecture.json
@@ -1,0 +1,1 @@
+{"kind": "none"}
